add knowledge pipeline manager
