+++ rootfs/knowledge_pipelines/pipelines.toml
@@ -1,0 +1,2 @@
+[[pipelines]]
+"local_photos_embedding"
