+++ src/knowledge/__init__.py
@@ -3,3 +3,4 @@ from .vector import *
 from .data import *
 from .store import KnowledgeStore
 from .core_object import *
+from .pipeline import *
