+++ src/knowledge/input/__init__.py
@@ -1,0 +1,1 @@
+import local_dir
